Bug 411080 - TableUtility: exportRowsAsCsv doesn't handle BigDecimal properly
https://bugs.eclipse.org/bugs/show_bug.cgi?id=411080

Fixes CSV export and adds JUnit tests.

Reviewed-on: https://git.eclipse.org/r/14095
Tested-by: Hudson CI
IP-Clean: Ken Lee <kle@bsiag.com>

+++ org.eclipse.scout.rt.client/src/org/eclipse/scout/rt/client/ui/basic/table/TableUtility.java
@@ -228,7 +228,7 @@ public final class TableUtility {
         format = ((ILongColumn) col).getFormat();
       }
       else if (col instanceof IBigDecimalColumn) {
-        type = Long.class;
+        type = Double.class;
         byValue = true;
         format = ((IBigDecimalColumn) col).getFormat();
       }
